Large refactoring for implementation using BasicStats. Added Mode View and accepts now Doubles instead of integers

+++ src/Controller.java
@@ -2,6 +2,6 @@
 public interface Controller {
 
 	public void addModel(Model model);
-	public void updateModels(Integer num);
+	public void updateModels(Double num);
 
 }
